refactor(sync-core): split cross-tab modules into many small focused files
Restructure the cross-tab/ directory so each file has one clear job, no
file exceeds ~320 lines, and the routing logic is composable from
independently-testable pieces.

Renames (filenames now match their primary exported function):

  presenter.ts        -> createPresenter.ts        (Presenter type co-located)
  leader.ts           -> createLeader.ts           (Leader type co-located)
  leader-router.ts    -> createLeaderRouter.ts     (LeaderRouter type co-located)
  cross-tab-socket.ts -> createCrossTabSocket.ts   (CrossTabSocket type +
                                                    Options co-located)

Splits:

  protocol.ts        -> types.ts               (shared types only)
                      + createCrossTabChannel.ts (the channel factory function)

  browser-context.ts -> types.ts               (BrowserContext interface,
                                                shared by Leader/Presenter/Glue)
                      + defaultBrowserContext.ts (the default value)

  leader-router.ts   -> createPushRouting.ts   (routing tables: register/
                                                consume push and connect,
                                                no I/O — pure state)
                      + routeServerMessage.ts  (stateless dispatch logic:
                                                takes routing + delivery
                                                callbacks, splits data
                                                batches, synthesizes patches)
                      + createLeaderRouter.ts  (slim glue: owns socket +
                                                channel, wires the above)

  cross-tab-socket.ts -> createFollowerReceiver.ts (follower-side channel
                                                    dispatch, extracted)
                       + createCrossTabSocket.ts   (slim glue: composes
                                                    all the pieces)

New pure unit tests:

  createPushRouting.test.ts    (8 tests — direct tests of the routing
                                tables without any channel setup)
  routeServerMessage.test.ts   (10 tests — synthesizePatchForSiblings +
                                routeServerMessage with a spy dispatch)

Integration tests split by topic, matching their primary module:

  createPresenter.test.ts      Presenter election (was inside the big file)
  createLeader.test.ts         Leader election + visibility-driven handoff
  createLeaderRouter.test.ts   Message routing + sibling-patch synthesis +
                               status propagation
  createCrossTabSocket.test.ts Fallback + cleanup + end-to-end glue

Shared test doubles live in `testDoubles.ts` so each test file can
import directly without reaching into a sibling. `createMockBrowserEnv`
renamed to `createMockBrowserContext` to match the production naming;
its return shape changes `{ env }` -> `{ ctx }` for the same reason.

Tiny pure helpers extracted within modules:

  isHigherClaim(claim, tabId, vsClaim, vsTabId)  in createPresenter.ts
  shouldWantLock(myIsVisible, peerVisibility)   in createLeader.ts

These move out of the factory closures to module scope, so they're
named, documented, and individually inspectable.

Net effect:

  - Biggest source file: 376 -> 320 lines (createCrossTabSocket.ts).
  - All other files: <= 215 lines.
  - 19 new pure unit tests covering the pieces that previously could
    only be exercised through the full CrossTabSocket stack.

No behavior change.
