+++ packages/sync-core/src/lib/cross-tab/createLeaderRouter.test.ts
@@ -1,315 +1,18 @@
 import { TLRecord } from '@tldraw/tlschema'
 import { beforeEach, describe, expect, it } from 'vitest'
 import { TLSocketClientSentEvent, TLSocketServerSentEvent } from '../protocol'
+import { TLPersistentClientSocketStatus } from '../TLSyncClient'
 import {
-	TLPersistentClientSocket,
-	TLPersistentClientSocketStatus,
-	TLSocketStatusChangeEvent,
-	TLSocketStatusListener,
-} from '../TLSyncClient'
-import { BrowserContext } from './browser-context'
-import { createCrossTabSocket } from './cross-tab-socket'
-import { BroadcastChannelLike, CrossTabLockManager } from './protocol'
+	buildTab,
+	createMockBrowserContext,
+	createMockChannelFactory,
+	createMockLockManager,
+	FakeSocket,
+	flushMicrotasks,
+} from './testDoubles'
+import { CrossTabLockManager } from './types'
 
-// =====================================================================
-// Test doubles
-// =====================================================================
-
-/**
- * In-process BroadcastChannel mock. All channels constructed via
- * `createMockChannel(name)` with the same name receive each other's messages
- * (and not their own).
- */
-function createMockChannelFactory() {
-	const channelsByName = new Map<string, Set<MockChannel>>()
-
-	class MockChannel implements BroadcastChannelLike {
-		private listeners = new Set<(ev: MessageEvent) => void>()
-		closed = false
-
-		constructor(public readonly name: string) {
-			const set = channelsByName.get(name) ?? new Set()
-			set.add(this)
-			channelsByName.set(name, set)
-		}
-
-		postMessage(msg: any) {
-			if (this.closed) return
-			const peers = channelsByName.get(this.name)
-			if (!peers) return
-			// Deliver synchronously to peers (excluding self), matching the
-			// browser's behavior of not echoing to the sender.
-			for (const peer of peers) {
-				if (peer === this) continue
-				if (peer.closed) continue
-				peer.listeners.forEach((cb) => cb({ data: msg } as MessageEvent))
-			}
-		}
-
-		addEventListener(_type: 'message', cb: (ev: MessageEvent) => void) {
-			this.listeners.add(cb)
-		}
-
-		removeEventListener(_type: 'message', cb: (ev: MessageEvent) => void) {
-			this.listeners.delete(cb)
-		}
-
-		close() {
-			this.closed = true
-			const set = channelsByName.get(this.name)
-			set?.delete(this)
-			this.listeners.clear()
-		}
-	}
-
-	return {
-		create: (name: string) => new MockChannel(name),
-	}
-}
-
-/**
- * Lock manager mock with FIFO queueing per lock name. Resolves the next
- * waiter's callback only after the current holder releases.
- */
-function createMockLockManager(): CrossTabLockManager & {
-	holders: Map<string, { resolve: () => void } | null>
-} {
-	const queues = new Map<string, Array<() => void>>()
-	const holders = new Map<string, { resolve: () => void } | null>()
-
-	const run = (name: string) => {
-		const q = queues.get(name)
-		if (!q || q.length === 0) {
-			holders.set(name, null)
-			return
-		}
-		const next = q.shift()!
-		next()
-	}
-
-	return {
-		holders,
-		async request(name, _options, callback) {
-			return new Promise<unknown>((outerResolve) => {
-				const start = () => {
-					const release = () => {
-						holders.set(name, null)
-						outerResolve(undefined)
-						run(name)
-					}
-					holders.set(name, { resolve: release })
-					// Per the Web Locks spec, the lock is held until the callback's
-					// promise resolves. We rely on CrossTabSocket to resolve the
-					// promise via its own release mechanism (close()).
-					callback().then(release, release)
-				}
-				if (holders.get(name)) {
-					const q = queues.get(name) ?? []
-					q.push(start)
-					queues.set(name, q)
-				} else {
-					start()
-				}
-			})
-		},
-	}
-}
-
-/**
- * A fake socket that records sent messages and lets tests push server messages
- * and status changes.
- */
-class FakeSocket
-	implements
-		TLPersistentClientSocket<TLSocketClientSentEvent<TLRecord>, TLSocketServerSentEvent<TLRecord>>
-{
-	connectionStatus: TLPersistentClientSocketStatus = 'offline'
-	sent: TLSocketClientSentEvent<TLRecord>[] = []
-	closed = false
-	private messageListeners = new Set<(msg: TLSocketServerSentEvent<TLRecord>) => void>()
-	private statusListeners = new Set<TLSocketStatusListener>()
-
-	sendMessage(msg: TLSocketClientSentEvent<TLRecord>) {
-		this.sent.push(msg)
-	}
-	onReceiveMessage(cb: (val: TLSocketServerSentEvent<TLRecord>) => void) {
-		this.messageListeners.add(cb)
-		return () => this.messageListeners.delete(cb)
-	}
-	onStatusChange(cb: TLSocketStatusListener) {
-		this.statusListeners.add(cb)
-		return () => this.statusListeners.delete(cb)
-	}
-	restart() {
-		// Mimic a reset: go offline then online again so that listeners react.
-		this._emitStatus({ status: 'offline' })
-	}
-	close() {
-		this.closed = true
-		this.messageListeners.clear()
-		this.statusListeners.clear()
-	}
-
-	// Test helpers
-	_emitStatus(ev: TLSocketStatusChangeEvent) {
-		this.connectionStatus = ev.status
-		this.statusListeners.forEach((cb) => cb(ev))
-	}
-	_emitServerMessage(msg: TLSocketServerSentEvent<TLRecord>) {
-		this.messageListeners.forEach((cb) => cb(msg))
-	}
-}
-
-// Convenience: build a tab and capture references to the socket factory
-// argument so the test can drive the underlying "WS".
-function buildTab(opts: {
-	channels: ReturnType<typeof createMockChannelFactory>
-	locks: CrossTabLockManager
-	channelKey: string
-	tabId: string
-	sockets: FakeSocket[]
-	browserContext?: BrowserContext | null
-}) {
-	const ct = createCrossTabSocket(() => 'ws://test', {
-		channelKey: opts.channelKey,
-		channel: opts.channels.create(`tldraw-room-${opts.channelKey}`),
-		locks: opts.locks,
-		tabId: opts.tabId,
-		browserContext: opts.browserContext,
-		createSocket: () => {
-			const s = new FakeSocket()
-			opts.sockets.push(s)
-			return s
-		},
-	})
-	return ct
-}
-
-/**
- * Mock browser env that lets the test push focus and visibility changes
- * deterministically. Starts unfocused and hidden by default so tests can opt
- * in to whatever initial state they want.
- */
-function createMockBrowserEnv(initial?: { focused?: boolean; visible?: boolean }) {
-	let focused = initial?.focused ?? false
-	let visible = initial?.visible ?? false
-	const focusListeners = new Set<() => void>()
-	const visibilityListeners = new Set<() => void>()
-
-	const env: BrowserContext = {
-		hasFocus: () => focused,
-		isVisible: () => visible,
-		onFocus(cb) {
-			focusListeners.add(cb)
-			return () => focusListeners.delete(cb)
-		},
-		onVisibilityChange(cb) {
-			visibilityListeners.add(cb)
-			return () => visibilityListeners.delete(cb)
-		},
-	}
-
-	return {
-		env,
-		focus() {
-			focused = true
-			focusListeners.forEach((cb) => cb())
-		},
-		blur() {
-			focused = false
-			// `blur` doesn't fire any handler — the design is "most recently
-			// focused tab stays presenter until another tab claims".
-		},
-		setVisible(value: boolean) {
-			visible = value
-			visibilityListeners.forEach((cb) => cb())
-		},
-	}
-}
-
-// Wait for queued microtasks. The lock manager and broadcast channel deliver
-// synchronously, but lock callbacks (which return promises) need a turn of the
-// microtask queue to settle.
-const flushMicrotasks = () => new Promise<void>((resolve) => setImmediate(resolve))
-
-// =====================================================================
-// Tests
-// =====================================================================
-
-describe('CrossTabSocket: leader election', () => {
-	let channels: ReturnType<typeof createMockChannelFactory>
-	let locks: CrossTabLockManager
-
-	beforeEach(() => {
-		channels = createMockChannelFactory()
-		locks = createMockLockManager()
-	})
-
-	it('the first tab to request the lock becomes leader', async () => {
-		const sockets: FakeSocket[] = []
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-		})
-		await flushMicrotasks()
-
-		expect(sockets).toHaveLength(1)
-		expect(a.mode).toBe('leader')
-		a.close()
-	})
-
-	it('subsequent tabs become followers', async () => {
-		const sockets: FakeSocket[] = []
-		const a = buildTab({ channels, locks, channelKey: 'room1', tabId: 'A', sockets })
-		await flushMicrotasks()
-		const b = buildTab({ channels, locks, channelKey: 'room1', tabId: 'B', sockets })
-		await flushMicrotasks()
-
-		// Only one underlying socket exists (the leader's).
-		expect(sockets).toHaveLength(1)
-		expect(a.mode).toBe('leader')
-		expect(b.mode).toBe('follower')
-		a.close()
-		b.close()
-	})
-
-	it('different rooms elect leaders independently', async () => {
-		const sockets: FakeSocket[] = []
-		const a = buildTab({ channels, locks, channelKey: 'room1', tabId: 'A', sockets })
-		const b = buildTab({ channels, locks, channelKey: 'room2', tabId: 'B', sockets })
-		await flushMicrotasks()
-
-		expect(a.mode).toBe('leader')
-		expect(b.mode).toBe('leader')
-		expect(sockets).toHaveLength(2)
-		a.close()
-		b.close()
-	})
-
-	it('when the leader closes, the next follower takes over', async () => {
-		const sockets: FakeSocket[] = []
-		const a = buildTab({ channels, locks, channelKey: 'room1', tabId: 'A', sockets })
-		await flushMicrotasks()
-		const b = buildTab({ channels, locks, channelKey: 'room1', tabId: 'B', sockets })
-		await flushMicrotasks()
-
-		expect(a.mode).toBe('leader')
-		expect(b.mode).toBe('follower')
-
-		a.close()
-		await flushMicrotasks()
-
-		expect(b.mode).toBe('leader')
-		// A new socket was opened by B.
-		expect(sockets).toHaveLength(2)
-		b.close()
-	})
-})
-
-describe('CrossTabSocket: message routing', () => {
+describe('createLeaderRouter: message routing', () => {
 	let channels: ReturnType<typeof createMockChannelFactory>
 	let locks: CrossTabLockManager
 	let sockets: FakeSocket[]
@@ -540,7 +243,7 @@ describe('CrossTabSocket: message routing', () => {
 	})
 })
 
-describe('CrossTabSocket: sibling-patch synthesis', () => {
+describe('createLeaderRouter: sibling-patch synthesis', () => {
 	let channels: ReturnType<typeof createMockChannelFactory>
 	let locks: CrossTabLockManager
 	let sockets: FakeSocket[]
@@ -827,7 +530,7 @@ describe('CrossTabSocket: sibling-patch synthesis', () => {
 	})
 })
 
-describe('CrossTabSocket: status propagation', () => {
+describe('createLeaderRouter: status propagation', () => {
 	let channels: ReturnType<typeof createMockChannelFactory>
 	let locks: CrossTabLockManager
 	let sockets: FakeSocket[]
@@ -875,368 +578,3 @@ describe('CrossTabSocket: status propagation', () => {
 	})
 })
 
-describe('CrossTabSocket: fallback mode', () => {
-	it('uses a per-tab socket when locks are unavailable', async () => {
-		const channels = createMockChannelFactory()
-		const sockets: FakeSocket[] = []
-		const a = buildTab({
-			channels,
-			locks: null as any, // null forces fallback
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-		})
-
-		expect(a.mode).toBe('fallback')
-		expect(sockets).toHaveLength(1)
-
-		// sendMessage goes straight to the underlying socket.
-		a.sendMessage({ type: 'ping' })
-		expect(sockets[0].sent).toEqual([{ type: 'ping' }])
-
-		a.close()
-	})
-
-	it('two fallback tabs each open their own socket', async () => {
-		const channels = createMockChannelFactory()
-		const sockets: FakeSocket[] = []
-		const a = buildTab({
-			channels,
-			locks: null as any,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-		})
-		const b = buildTab({
-			channels,
-			locks: null as any,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-		})
-
-		expect(sockets).toHaveLength(2)
-		expect(a.mode).toBe('fallback')
-		expect(b.mode).toBe('fallback')
-
-		a.close()
-		b.close()
-	})
-})
-
-describe('CrossTabSocket: presenter election', () => {
-	let channels: ReturnType<typeof createMockChannelFactory>
-	let locks: CrossTabLockManager
-	let sockets: FakeSocket[]
-
-	beforeEach(() => {
-		channels = createMockChannelFactory()
-		locks = createMockLockManager()
-		sockets = []
-	})
-
-	it('the tab focused at construction is the initial presenter', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-
-		expect(a.$isPresenter.get()).toBe(true)
-		a.close()
-	})
-
-	it('a tab without focus at construction is not the presenter', async () => {
-		const aEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-
-		expect(a.$isPresenter.get()).toBe(false)
-		a.close()
-	})
-
-	it('a focus event makes that tab the presenter and demotes the previous one', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const bEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-		const b = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-			browserContext: bEnv.env,
-		})
-		await flushMicrotasks()
-
-		// A focused at construction; B did not.
-		expect(a.$isPresenter.get()).toBe(true)
-		expect(b.$isPresenter.get()).toBe(false)
-
-		// Now focus B.
-		bEnv.focus()
-
-		expect(a.$isPresenter.get()).toBe(false)
-		expect(b.$isPresenter.get()).toBe(true)
-
-		a.close()
-		b.close()
-	})
-
-	it('the last-focused tab wins ties between simultaneous claims', async () => {
-		const aEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const bEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-		const b = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-			browserContext: bEnv.env,
-		})
-		await flushMicrotasks()
-
-		// Both start unfocused — neither is presenter.
-		expect(a.$isPresenter.get()).toBe(false)
-		expect(b.$isPresenter.get()).toBe(false)
-
-		// A focuses first.
-		aEnv.focus()
-		expect(a.$isPresenter.get()).toBe(true)
-		expect(b.$isPresenter.get()).toBe(false)
-
-		// Then B focuses — takes over.
-		bEnv.focus()
-		expect(a.$isPresenter.get()).toBe(false)
-		expect(b.$isPresenter.get()).toBe(true)
-
-		// Then A focuses again — takes back.
-		aEnv.focus()
-		expect(a.$isPresenter.get()).toBe(true)
-		expect(b.$isPresenter.get()).toBe(false)
-
-		a.close()
-		b.close()
-	})
-
-	it('blur alone does not demote the presenter — it stays until another tab claims', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-
-		expect(a.$isPresenter.get()).toBe(true)
-
-		// Lose focus to something outside our pool (the OS taskbar, another app).
-		aEnv.blur()
-
-		// Still presenter — nothing has claimed.
-		expect(a.$isPresenter.get()).toBe(true)
-
-		a.close()
-	})
-
-	it('fallback-mode tabs are always presenters', async () => {
-		const a = buildTab({
-			channels,
-			locks: null as any,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-		})
-		await flushMicrotasks()
-
-		expect(a.$isPresenter.get()).toBe(true)
-		a.close()
-	})
-})
-
-describe('CrossTabSocket: visibility-driven leader handoff', () => {
-	let channels: ReturnType<typeof createMockChannelFactory>
-	let locks: CrossTabLockManager & { holders: Map<string, { resolve: () => void } | null> }
-	let sockets: FakeSocket[]
-
-	beforeEach(() => {
-		channels = createMockChannelFactory()
-		locks = createMockLockManager()
-		sockets = []
-	})
-
-	it('the leader releases the lock when it becomes hidden and another tab is visible', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const bEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-		const b = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-			browserContext: bEnv.env,
-		})
-		await flushMicrotasks()
-
-		expect(a.mode).toBe('leader')
-		expect(b.mode).toBe('follower')
-
-		// A becomes hidden. B is still visible.
-		aEnv.setVisible(false)
-		await flushMicrotasks()
-
-		// Lock should have migrated to B.
-		expect(b.mode).toBe('leader')
-		// Two sockets have been opened over the lifetime of the test.
-		expect(sockets).toHaveLength(2)
-
-		a.close()
-		b.close()
-	})
-
-	it('a hidden tab keeps the lock if no other tab is visible', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const bEnv = createMockBrowserEnv({ focused: false, visible: false })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-		const b = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-			browserContext: bEnv.env,
-		})
-		await flushMicrotasks()
-
-		expect(a.mode).toBe('leader')
-
-		// A becomes hidden — but B is also hidden, so A should hang onto the lock.
-		aEnv.setVisible(false)
-		await flushMicrotasks()
-
-		expect(a.mode).toBe('leader')
-		expect(sockets).toHaveLength(1)
-
-		a.close()
-		b.close()
-	})
-
-	it('a previously-hidden tab re-acquires the lock when it becomes visible again', async () => {
-		const aEnv = createMockBrowserEnv({ focused: true, visible: true })
-		const bEnv = createMockBrowserEnv({ focused: false, visible: true })
-		const a = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'A',
-			sockets,
-			browserContext: aEnv.env,
-		})
-		await flushMicrotasks()
-		const b = buildTab({
-			channels,
-			locks,
-			channelKey: 'room1',
-			tabId: 'B',
-			sockets,
-			browserContext: bEnv.env,
-		})
-		await flushMicrotasks()
-
-		// A is leader. A becomes hidden — hands off to B.
-		aEnv.setVisible(false)
-		await flushMicrotasks()
-		expect(b.mode).toBe('leader')
-
-		// A becomes visible again. It should request the lock; B keeps it for
-		// now (B is still visible), so A waits in the queue. If B then becomes
-		// hidden, A would take over.
-		aEnv.setVisible(true)
-		await flushMicrotasks()
-		expect(b.mode).toBe('leader')
-
-		bEnv.setVisible(false)
-		await flushMicrotasks()
-		expect(a.mode).toBe('leader')
-
-		a.close()
-		b.close()
-	})
-})
-
-describe('CrossTabSocket: cleanup', () => {
-	it('close() releases the lock', async () => {
-		const channels = createMockChannelFactory()
-		const locks = createMockLockManager() as any
-		const sockets: FakeSocket[] = []
-
-		const a = buildTab({ channels, locks, channelKey: 'room1', tabId: 'A', sockets })
-		await flushMicrotasks()
-
-		expect(locks.holders.get('tldraw-leader-room1')).not.toBeNull()
-		a.close()
-		await flushMicrotasks()
-		expect(locks.holders.get('tldraw-leader-room1')).toBeNull()
-	})
-
-	it('after close(), sendMessage throws', async () => {
-		const channels = createMockChannelFactory()
-		const locks = createMockLockManager()
-		const sockets: FakeSocket[] = []
-
-		const a = buildTab({ channels, locks, channelKey: 'room1', tabId: 'A', sockets })
-		await flushMicrotasks()
-		a.close()
-
-		expect(() => a.sendMessage({ type: 'ping' })).toThrow()
-	})
-})
